feat(logic): refactored code and added welcome function logic
Moved game state functions from `game/__init__.py` to `game/logic.py`.
Updated `run` function to call the game state actions from `Logic` class.
Added initial logic of `WELCOME` which sets the player and enemy objects.
